+++ tests/testthat/setup-shinytest2.R
@@ -6,6 +6,13 @@ open_collapse <- function(id, value) {
   return(runstring)
 }
 
+testthat::local_edition(3)
+
+# announce snapshots
+for (fpath in Sys.glob(file.path(testthat::test_path(), "_snaps", "*", "*"))) {
+  testthat::announce_snapshot_file(fpath)
+}
+
 if (Sys.getenv("MAP_SHINYTEST") == 1) {
   project_root = file.path(testthat::test_path(), "../..")
   
@@ -127,3 +134,38 @@ if (Sys.getenv("MAP_SHINYTEST") == 1) {
     }
   )
 }
+
+# for testing plotly plots, see https://rdrr.io/cran/plotly/src/tests/testthat/helper-vdiffr.R
+visual_testing = TRUE
+
+imageServer <- if (visual_testing) {
+  # https://github.com/plotly/plotly.R/issues/2179
+  reticulate::py_run_string("import sys") 
+  kaleido() 
+} else {
+  list(transform = function(...) stop("Visual testing is disabled!"))
+}
+
+# define logic for writing svg
+write_plotly_svg <- function(p, file, title = "") {
+  # before exporting, specify trace[i].uid so resulting svg is deterministic
+  # https://github.com/plotly/orca/issues/133
+  p <- plotly_build(p)
+  uid_data <- paste0("-vdiffr-plotly-", seq_along(p$x$data))
+  p$x$data <- Map(function(tr, id) { tr$uid <- id; tr }, p$x$data, uid_data)
+  
+  # write svg to disk
+  owd <- setwd(dirname(file))
+  on.exit(setwd(owd))
+  imageServer$transform(p, file = basename(file), width = 640, height = 480)
+  
+  # strip out non-deterministic fullLayout.uid
+  # TODO: if and when plotly provides an API to pre-specify, use it!
+  svg_txt <- readLines(file, warn = FALSE)
+  strextract <- function(str, pattern) regmatches(str, regexpr(pattern, str))
+  def <- strextract(svg_txt, 'defs id=\\"defs-[[:alnum:]]+\\"')
+  uid <- sub("defs-", "", strextract(def, "defs-[[:alnum:]]+"))
+  svg_txt <- gsub(uid, "", svg_txt, fixed = TRUE)
+  writeLines(svg_txt, file)
+}
+